requirements.txt
PySide6>=6.4.0
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0  # For async requests
python-dateutil>=2.8.2
google-api-python-client>=2.0.0  # For Google Custom Search API
